requirements.txt
selenium
pytest
requests
allure-pytest
faker
